renamed file and corrections
Figured out the sign convention problem in the original version and now note the convention

interaction point solver works for all vertical slices of P where the interaction surface exists in all 4 quadrants. Some instances where the solution window for the solver is too small and won't converge on the correct angle seems to be limited to tension checks.