Attempt to handle skewed tensile zones
attempt to handle skewed tensile zones by first shifting the diagram so the point of maximum tension lies at 0,0 and then running the solver for the interaction point checking against the shifted goal P and angle. Seems to work for the most part but doesn't capture the point of 0 moment under tension yet.