skewed tensile zone stuff didn't pan so removed it
after further testing the skewed tensile method I was trying didn't work think what needs to be is a full 3d transformation of the line from Pn+ to Pn- to align with the z-axis and perform checks against that.

worksheet is slowww because I am letting the cell formulas do all the work so the sheet executes in full on each iteration of any of the methods. This is really meant as a sandbox for analysis ideas, anticipate once translated to full Python or VBA should see significant speed boosts.